### 2019-06-08 #1496 @whedon
[PRE REVIEW]: CGIMP: Real-time exploration and covariate projection for self-organizing map datasets'
**Submitting author:** @adadiehl (<a href="http://orcid.org/0000-0002-0915-4570">Adam Diehl</a>)
**Repository:** <a href="https://github.com/Boyle-Lab/CGIMP" target ="_blank">https://github.com/Boyle-Lab/CGIMP</a>
**Version:** 1.0.0
