### 2019-06-08 #1496 @labarba
@adadiehl — You will need to rename the file here as `paper.md`
https://github.com/Boyle-Lab/CGIMP/tree/master/paper
