### 2019-06-10 #1496 @adadiehl
I have renamed the manuscript as paper.md. Thank you!

-Adam
