### 2019-06-11 53b8818
fixed typo in paper.md header
