### 2019-06-11 #1496 @adadiehl
You were correct. I've corrected it.

On Tue, Jun 11, 2019 at 2:11 PM Lorena Pantano <notifications@github.com>
wrote:

> @adadiehl <https://github.com/adadiehl>, it is possible you have a typo
> in the header for autors instead of authors?
>
