### 2019-06-13 #1496 @lpantano
can you make sure: name, orcid and affiliation have the same indentation please?
