### 2019-06-13 #1496 @adadiehl
On Thu, Jun 13, 2019 at 1:27 PM Lorena Pantano <notifications@github.com>
wrote:

> can you make sure: name, orcid and affiliation have the same indentation
> please?
>
