### 2019-06-13 fa33137
fixed missing , in CGIMP_manuscript.bib
