+++ paper/CGIMP_manuscript.bib
@@ -5,7 +5,7 @@
 }
 
 @article{Wehrens:2007,
-        author = {{Wehrens}, R and {Buydens}, LMC}
+        author = {{Wehrens}, R and {Buydens}, LMC},
 	year = {2007},
         title = {{Self-and super-organizing maps in R: the Kohonen package}},
         journal = {Journal of Statistical Software}
