### 2019-06-13 9b8ac1d
Fixed broken refs in paper.md
