### 2019-06-13 #1496 @adadiehl
All appears to be well with the proof now.
