### 2019-06-13 0e8ad51
affiliation 2 not showing up for AP Boyle. Fix attempt.
