### 2019-06-13 b30be71
affiliation 2 not showing up for AP Boyle. Fix attempt.
