### 2019-06-13 eef4008
affiliation 2 not showing up for AP Boyle. Fix attempt.
