### 2019-06-14 #1496 @adadiehl
It looks like pdf generation may have stalled with the most recent version,
which corrects a problem with the author affiliations for AP Boyle. As far
as I know, this should be the last correction that needs to be made.
