### 2019-06-21 #1496 @lpantano
@adadiehl can you check the references and add the DIO to them as shown here: https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography

Thanks
