### 2019-06-21 26d7827
Added ability to choose map fill color
