### 2019-06-21 5b80bc5
Added ability to choose map fill color
