### 2019-06-21 82f5aad
Added dois to bibliography
