+++ paper/CGIMP_manuscript.bib
@@ -8,7 +8,8 @@
         author = {{Wehrens}, R and {Buydens}, LMC},
 	year = {2007},
         title = {{Self-and super-organizing maps in R: the Kohonen package}},
-        journal = {Journal of Statistical Software}
+        journal = {Journal of Statistical Software},
+	doi = {10.18637/jss.v021.i05}
 }
 
 @article{Xie:2013,
@@ -19,7 +20,8 @@
         volume = {155},
         number = {3},
         pages = {713--724},
-        month = oct
+        month = oct,
+	doi = {10.1016/j.cell.2013.09.043}
 }
 
 @article{Boyle:2014,
@@ -30,7 +32,8 @@
         volume = {512},
         number = {7515},
         pages = {453--456},
-        month = aug
+        month = aug,
+	doi = {10.1038/nature13668}
 }
 
 @article{Kohonen:1990,
@@ -40,7 +43,8 @@
         year = {1990},
         volume = {78},
         number = {9},
-        pages = {1464--1480}
+        pages = {1464--1480},
+	doi = {10.1109/5.58325}
 }
 
 @article{Quinlan:2010,
@@ -51,7 +55,8 @@
         volume = {26},
         number = {6},
         pages = {841--842},
-        month = mar
+        month = mar,
+	doi = {10.1093/bioinformatics/btq033}
 }
 
 @article{Diehl:2018,
@@ -61,5 +66,6 @@
         year = {2018},
         volume = {46},
         number = {4},
-        pages = {1878--1894}
+        pages = {1878--1894},
+	doi = {10.1093/nar/gky018}
 }
