### 2019-06-21 #1496 @adadiehl
I’ve fixed DOIs for all references that I can. One does not appear to have a DOI. How do I handle that reference?

"The wind of heaven is that which flows between a horse's ears."

> On Jun 21, 2019, at 1:42 PM, Lorena Pantano <notifications@github.com> wrote:
>
> @adadiehl can you check the references and add the DIO to them as shown here: https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography
>
> Thanks
>
