### 2019-06-24 #1520 @whedon
[REVIEW]: CGIMP: Real-time exploration and covariate projection for self-organizing map datasets'
**Submitting author:** @adadiehl (<a href="http://orcid.org/0000-0002-0915-4570">Adam Diehl</a>)
**Repository:** <a href="https://github.com/Boyle-Lab/CGIMP" target ="_blank">https://github.com/Boyle-Lab/CGIMP</a>
**Version:** 1.0.0
**Reviewers:** @adriancbondia, @ @arunhpatil, @arunhpatil

@adriancbondia & @ @arunhpatil, please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
