### 2019-06-25 eeba21b
Added ability to fix density scale from 0 to 1.
