### 2019-06-25 ae64cc1
Added JOSS badge.
