### 2019-06-26 1bf9bd4
Added pre-query to get numeric facet ranges and switched numeric components to RangeInput components. (Reduces number of initial queries and improves usability)
