### 2019-06-26 f254dcf
Added spinner dialog to indicate visually when data and search components are fully loaded and ready for user interactions. Should reduce the chance of ending up with wrong data due to async request "races" and improve user experience.
