### 2019-06-27 #1520 @lpantano
Hi @adadiehl,

Do you have a link for the reference with DOI, or what exactly is that, a chapter in a book, a talk...?
