### 2019-06-27 #1520 @adadiehl
It is a master's thesis in an online repository. I can include a link.

On Thu, Jun 27, 2019 at 9:54 AM Lorena Pantano <notifications@github.com>
wrote:

> Hi @adadiehl <https://github.com/adadiehl>,
>
> Do you have a link for the reference with DOI, or what exactly is that, a
> chapter in a book, a talk...?
>
