### 2019-06-27 4690ba2
Added URL for Yuan et al.
