+++ paper/CGIMP_manuscript.bib
@@ -1,7 +1,8 @@
 @article{Yuan:2018,
         author = {{Yuan}, L},
         title = {{Implementation of Self-Organizing Maps with Python}},
-        year = {2018}
+        year = {2018},
+	url = {https://digitalcommons.uri.edu/theses/1244/}
 }
 
 @article{Wehrens:2007,
