### 2019-06-27 #1520 @lpantano
I think you can make the change in the .bib file to have it like this:
https://github.com/arfon/fidgit/blob/master/paper/paper.bib#L17

On June 27, 2019 at 10:26:40 AM, Adam Diehl (notifications@github.com)
wrote:

It is a master's thesis in an online repository. I can include a link.

On Thu, Jun 27, 2019 at 9:54 AM Lorena Pantano <notifications@github.com>
wrote:

> Hi @adadiehl <https://github.com/adadiehl>,
>
> Do you have a link for the reference with DOI, or what exactly is that, a
> chapter in a book, a talk...?
>
