### 2019-06-27 4b24892
Added urldate for Yuan et al.
