+++ paper/CGIMP_manuscript.bib
@@ -2,7 +2,8 @@
         author = {{Yuan}, L},
         title = {{Implementation of Self-Organizing Maps with Python}},
         year = {2018},
-	url = {https://digitalcommons.uri.edu/theses/1244/}
+	url = {https://digitalcommons.uri.edu/theses/1244/},
+	urldate = {2019-06-27}
 }
 
 @article{Wehrens:2007,
