### 2019-06-27 ad1d9dd
Corrected reference type for Yuan et al.
