+++ paper/CGIMP_manuscript.bib
@@ -1,4 +1,4 @@
-@article{Yuan:2018,
+@online{Yuan:2018,
         author = {{Yuan}, L},
         title = {{Implementation of Self-Organizing Maps with Python}},
         year = {2018},
