### 2019-06-27 #1520 @adadiehl
Okay, I've made the changes. I think we're good to go.

On Thu, Jun 27, 2019 at 10:37 AM Lorena Pantano <notifications@github.com>
wrote:

> I think you can make the change in the .bib file to have it like this:
> https://github.com/arfon/fidgit/blob/master/paper/paper.bib#L17
>
>
>
> On June 27, 2019 at 10:26:40 AM, Adam Diehl (notifications@github.com)
> wrote:
>
> It is a master's thesis in an online repository. I can include a link.
>
> On Thu, Jun 27, 2019 at 9:54 AM Lorena Pantano <notifications@github.com>
> wrote:
>
> > Hi @adadiehl <https://github.com/adadiehl>,
> >
> > Do you have a link for the reference with DOI, or what exactly is that, a
> > chapter in a book, a talk...?
> >
> > —
> > You are receiving this because you were mentioned.
> > Reply to this email directly, view it on GitHub
> > <
>
> https://github.com/openjournals/joss-reviews/issues/1520?email_source=notifications&email_token=ACUMZSOXWFM4CPP6MVNGIS3P4TBBRA5CNFSM4H3A2IQ2YY3PNVWWK3TUL52HS4DFVREXG43VMVBW63LNMVXHJKTDN5WW2ZLOORPWSZGODYXGDSQ#issuecomment-506356170
> >,
> > or mute the thread
> > <
>
> https://github.com/notifications/unsubscribe-auth/ACUMZSP2LSURVDVLUIAOK53P4TBBRANCNFSM4H3A2IQQ
> >
> > .
> >
>
