### 2019-06-27 04e0896
Eliminated warning about nested div inside <p>
