+++ client/src/LoadAlert.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import Dialog from '@material-ui/core/Dialog';
 import DialogContent from '@material-ui/core/DialogContent';
-import DialogContentText from '@material-ui/core/DialogContentText';
 import DialogTitle from '@material-ui/core/DialogTitle';
 import CircularProgress from '@material-ui/core/CircularProgress';
 
@@ -45,9 +44,7 @@ class LoadAlertDialog extends React.Component {
 		        {"Working..."}
 	            </DialogTitle>
                     <DialogContent>
-                        <DialogContentText id="alert-dialog-description">
-		            <CircularProgress />
-                        </DialogContentText>
+		        <CircularProgress />
                     </DialogContent>
 	        </Dialog>
 	    </div>
