### 2019-07-08 #1520 @lpantano
Hi @adriancbondia  and @arunhpatil , I think you finished the review but can you confirm here to make it clear? thanks!
