### 2019-07-09 #1520 @adriancbondia
Sorry for the delay, yes, I can confirm it

El lun., 8 jul. 2019 15:43, Lorena Pantano <notifications@github.com>
escribió:

> Hi @adriancbondia <https://github.com/adriancbondia> and @arunhpatil
> <https://github.com/arunhpatil> , I think you finished the review but can
> you confirm here to make it clear? thanks!
>
