### 2019-07-09 #1520 @arunhpatil
Yes. The review is complete. The application CGIMP has an interactive browser with lot of features for users to explore, however as mentioned in the documentation, the first few attempts load slowly due to indexing which upon subsequent usage loads faster. This is a minor issue and can be further improvised. I confirm.
