### 2019-07-09 #1520 @adadiehl
Done!

https://doi.org/10.5281/zenodo.3276069

On Tue, Jul 9, 2019 at 9:06 AM Lorena Pantano <notifications@github.com>
wrote:

> Hi @adadiehl <https://github.com/adadiehl> , can you create a Zenodo
> archive where the title and authors match the paper version and give me
> back the doi link?
>
