### 2019-07-09 #1520 @adadiehl
I will make the fix. Just to clarify, you want the title on the archive to
match the manuscript title?

On Tue, Jul 9, 2019 at 11:43 AM Lorena Pantano <notifications@github.com>
wrote:

> Thanks, can you make sure the title is the same? I don't know if you would
> need a new doi link then, but let me know. Cheers
>
