### 2019-07-09 a7ceca5
Added updateParentState prop to Facets.js and made list keys more robust.
